requirements.txt
discord.py==2.3.2
SQLAlchemy==2.0.20
aiosqlite==0.19.0
